refactor(omnix-cli): Keep only exports in mod.rs
