+++ crates/omnix-cli/src/command/mod.rs
@@ -1,38 +1,6 @@
 pub mod ci;
 pub mod completion;
+pub mod core;
 pub mod health;
 pub mod init;
 pub mod show;
-
-use clap::Subcommand;
-use clap_verbosity_flag::{InfoLevel, Verbosity};
-
-#[derive(Subcommand, Debug)]
-pub enum Command {
-    Show(show::ShowCommand),
-
-    Init(init::InitCommand),
-
-    CI(ci::CICommand),
-
-    Health(health::HealthCommand),
-
-    Completion(completion::CompletionCommand),
-}
-
-impl Command {
-    pub async fn run(&self, verbosity: Verbosity<InfoLevel>) -> anyhow::Result<()> {
-        if !matches!(self, Command::Completion(_)) && !omnix_common::check::nix_installed() {
-            tracing::error!("Nix is not installed: https://nixos.asia/en/install");
-            std::process::exit(1);
-        }
-
-        match self {
-            Command::Show(cmd) => cmd.run().await,
-            Command::Init(cmd) => cmd.run().await,
-            Command::CI(cmd) => cmd.run(verbosity).await,
-            Command::Health(cmd) => cmd.run().await,
-            Command::Completion(cmd) => cmd.run(),
-        }
-    }
-}
